# server/go/go/model_employee_info_schema.go
/*
 * SPAN Tenure
 *
 * This is a simple API which is used to as a demo for using OpenAPI to define specifications.
 *
 * API version: 0.0.1
 * Contact: jayan.smart@spandigital.com
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type EmployeeInfoSchema struct {
	// The id of the employee
	Id int32 `json:"id,omitempty"`
	// The employee name
	Name string `json:"name,omitempty"`
	// The tenure of an employee in days
	Tenure int32 `json:"tenure,omitempty"`
}