# server/go/go/model_server_error.go
/*
 * SPAN Tenure
 *
 * This is a simple API which is used to as a demo for using OpenAPI to define specifications.
 *
 * API version: 0.0.1
 * Contact: jayan.smart@spandigital.com
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type ServerError struct {
	// Information about the error which has occurred.
	Response string `json:"response,omitempty"`
}